+++ app/shows/reports/guest_scorekeeper.py
@@ -5,6 +5,8 @@
 # vim: set noai syntax=python ts=4 sw=4:
 """WWDTM Show Guest Scorekeeper Report Functions."""
 
+from typing import Any
+
 from mysql.connector.connection import MySQLConnection
 from mysql.connector.pooling import PooledMySQLConnection
 
@@ -13,6 +15,82 @@ from .show_details import (
     retrieve_show_guests,
     retrieve_show_panelists,
 )
+
+
+def retrieve_best_of_shows_guest_scorekeeper(
+    database_connection: MySQLConnection | PooledMySQLConnection,
+) -> list[dict[str, Any]]:
+    """Retrieve a list of all Best Of shows with a guest scorekeeper.
+
+    Basic information for each show includes location, host,
+    scorekeeper, panelists and guest.
+    """
+    if not database_connection.is_connected():
+        database_connection.reconnect()
+
+    query = """
+        SELECT s.showid, s.showdate, s.bestof, s.repeatshowid, h.host,
+        h.hostslug, hm.guest AS host_guest, sk.scorekeeper,
+        sk.scorekeeperslug, l.venue, l.city, l.state
+        FROM ww_showhostmap hm
+        JOIN ww_hosts h ON h.hostid = hm.hostid
+        JOIN ww_showskmap skm ON skm.showid = hm.showid
+        JOIN ww_scorekeepers sk ON sk.scorekeeperid = skm.scorekeeperid
+        JOIN ww_shows s ON s.showid = hm.showid
+        JOIN ww_showlocationmap lm ON lm.showid = hm.showid
+        JOIN ww_locations l ON l.locationid = lm.locationid
+        WHERE s.showdate < NOW()
+        AND s.bestof = 1 AND skm.guest = 1
+        ORDER BY s.showdate ASC;
+    """
+    cursor = database_connection.cursor(dictionary=True)
+    cursor.execute(query)
+    results = cursor.fetchall()
+    cursor.close()
+
+    if not results:
+        return None
+
+    shows = []
+    show_count = 1
+    for row in results:
+        shows.append(
+            {
+                "count": show_count,
+                "id": row["showid"],
+                "date": row["showdate"],
+                "best_of": bool(row["bestof"]),
+                "repeat": bool(row["repeatshowid"]),
+                "original_show_date": (
+                    retrieve_show_date_by_id(
+                        show_id=row["repeatshowid"],
+                        database_connection=database_connection,
+                    )
+                    if row["repeatshowid"]
+                    else None
+                ),
+                "location": {
+                    "venue": row["venue"],
+                    "city": row["city"],
+                    "state": row["state"],
+                },
+                "host": row["host"],
+                "host_slug": row["hostslug"],
+                "host_guest": bool(row["host_guest"]),
+                "scorekeeper": row["scorekeeper"],
+                "scorekeeper_slug": row["scorekeeperslug"],
+                "panelists": retrieve_show_panelists(
+                    show_id=row["showid"], database_connection=database_connection
+                ),
+                "guests": retrieve_show_guests(
+                    show_id=row["showid"], database_connection=database_connection
+                ),
+            }
+        )
+
+        show_count += 1
+
+    return shows
 
 
 def retrieve_shows_guest_scorekeeper(